move request course form to separate component
